% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadRedcapReport.R
\name{downloadRedcapReport}
\alias{downloadRedcapReport}
\title{Download a REDCap report as a tibble}
\usage{
downloadRedcapReport(redcapTokenName, redcapUrl, redcapReportId)
}
\arguments{
\item{redcapTokenName}{The name of the environment variable holding the REDCap API token}

\item{redcapUrl}{The base URL of the REDCap server}

\item{redcapReportId}{The ID of the REDCap report to download}
}
\value{
A tibble containing the REDCap report data
}
\description{
Download a REDCap report as a tibble
}
\examples{
redcapTokenName <- "TOKEN"
redcapUrl <- "https://redcap.emory.edu/api/"
redcapReportId <- "46524"
report_data <- downloadRedcapReport(redcapTokenName, redcapUrl, redcapReportId)

}
