% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimumN.R
\name{minimumN}
\alias{minimumN}
\title{Calculate minimum sample size for a t-test}
\usage{
minimumN(x1, x2 = NULL)
}
\arguments{
\item{x1}{A vector of preliminary data (required)}

\item{x2}{An optional second vector of preliminary data}
}
\value{
The minimum sample size needed for a t-test with 0.8 power at alpha=0.05
}
\description{
Calculate minimum sample size for a t-test
}
\examples{
x1 = rnorm(100)
minimumN(x1)
x2 = rnorm(150)
minimumN(x1, x2)

}
