% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcApprox.R
\name{pcApprox}
\alias{pcApprox}
\title{Approximate data based on a specified number of principal components}
\usage{
pcApprox(x, npc)
}
\arguments{
\item{x}{The original data}

\item{npc}{The number of principal components to use for approximation}
}
\value{
The approximated data
}
\description{
Approximate data based on a specified number of principal components
}
\examples{
x = matrix(rnorm(50), nrow = 10)
approx_x = pcApprox(x, npc = 3)

}
