% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeNames.R
\name{standardizeNames}
\alias{standardizeNames}
\title{Standardize variable names in a tibble to "small_camel" case}
\usage{
standardizeNames(data)
}
\arguments{
\item{data}{A tibble}
}
\value{
The input tibble with variable names in "small_camel" case
}
\description{
Standardize variable names in a tibble to "small_camel" case
}
\examples{
data = tibble(MyVar = 1:5, 'Another Var' = 6:10)
standardizeNames(data)

}
