% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survCurv.R
\name{survCurv}
\alias{survCurv}
\title{Plot survival curve}
\usage{
survCurv(status, time)
}
\arguments{
\item{status}{A vector indicating the status of each individual (1 for event occurrence, 0 for censoring).}

\item{time}{A vector of time points corresponding to each individual's event or censoring time.}
}
\value{
None (plot is generated).
}
\description{
This function plots the survival curve based on the input status and time data.
}
\examples{
data = read.csv("https://jlucasmckay.bmi.emory.edu/global/bmi510/Labs-Materials/survival.csv")
survCurv(data$status, data$time)

}
