% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLikBernoulli.R
\name{logLikBernoulli}
\alias{logLikBernoulli}
\title{Calculate maximum likelihood estimate of p for Bernoulli data}
\usage{
logLikBernoulli(data)
}
\arguments{
\item{data}{A vector of 0s and 1s}
}
\value{
The value of p that maximizes the log-likelihood
}
\description{
This function calculates the maximum likelihood estimate of the parameter p
for Bernoulli data, which is represented as a vector of 0s and 1s.
}
\examples{
# logLikBernoulli(c(1,0,0,0,1,1,1))
}
