% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survCurv.R
\name{survCurv}
\alias{survCurv}
\title{Plot survival curve}
\usage{
survCurv(status, time)
}
\arguments{
\item{status}{A vector indicating the status (1 for event, 0 for censored).}

\item{time}{A vector of time points.}
}
\value{
None (plot is generated).
}
\description{
This function plots the survival curve based on the input status and time data.
}
\examples{
status = c(1, 1, 0, 1, 0)
time = c(10, 15, 20, 25, 30)
survCurv(status, time)

}
